Refactor karva_test
