+++ crates/karva_test/src/context/test_context.rs
@@ -9,36 +9,7 @@ use anyhow::Context;
 use insta::{Settings, internals::SettingsBindDropGuard};
 use tempfile::TempDir;
 
-/// Find the karva wheel in the target/wheels directory.
-/// Returns the path to the wheel file.
-pub fn find_karva_wheel() -> anyhow::Result<PathBuf> {
-    let karva_root = Path::new(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .and_then(|p| p.parent())
-        .ok_or_else(|| anyhow::anyhow!("Could not determine KARVA_ROOT"))?
-        .to_path_buf();
-
-    let wheels_dir = karva_root.join("target").join("wheels");
-
-    let entries = std::fs::read_dir(&wheels_dir)
-        .with_context(|| format!("Could not read wheels directory: {}", wheels_dir.display()))?;
-
-    for entry in entries {
-        let entry = entry?;
-        let file_name = entry.file_name();
-        if let Some(name) = file_name.to_str() {
-            if name.starts_with("karva-")
-                && std::path::Path::new(name)
-                    .extension()
-                    .is_some_and(|ext| ext.eq_ignore_ascii_case("whl"))
-            {
-                return Ok(entry.path());
-            }
-        }
-    }
-
-    anyhow::bail!("Could not find karva wheel in target/wheels directory");
-}
+use crate::{find_karva_wheel, utils::tempdir_filter};
 
 pub struct TestContext {
     _temp_dir: TempDir,
@@ -156,13 +127,13 @@ impl TestContext {
     }
 
     pub fn with_files<'a>(files: impl IntoIterator<Item = (&'a str, &'a str)>) -> Self {
-        let mut case = Self::new();
+        let mut case = Self::default();
         case.write_files(files).unwrap();
         case
     }
 
     pub fn with_file(path: impl AsRef<Path>, content: &str) -> Self {
-        let mut case = Self::new();
+        let mut case = Self::default();
         case.write_file(path, content).unwrap();
         case
     }
@@ -221,91 +192,10 @@ impl TestContext {
     pub fn mapped_path(&self, path: &str) -> Option<&PathBuf> {
         self.mapped_paths.get(path)
     }
-
-    #[must_use]
-    pub fn relative_path(&self, path: &Path) -> PathBuf {
-        PathBuf::from(path.strip_prefix(self.cwd()).unwrap())
-    }
 }
 
 impl Default for TestContext {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-fn tempdir_filter(path: &Path) -> String {
-    format!(r"{}\\?/?", regex::escape(path.to_str().unwrap()))
-}
-
-pub struct IntegrationTestContext {
-    test_env: TestContext,
-}
-
-impl Default for IntegrationTestContext {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl IntegrationTestContext {
-    #[must_use]
-    pub fn new() -> Self {
-        let test_env = TestContext::new();
-
-        Self { test_env }
-    }
-
-    #[must_use]
-    pub fn karva_bin(&self) -> PathBuf {
-        let venv_bin =
-            self.test_env
-                .cwd()
-                .join(".venv")
-                .join(if cfg!(windows) { "Scripts" } else { "bin" });
-        venv_bin.join(if cfg!(windows) { "karva.exe" } else { "karva" })
-    }
-
-    pub fn with_files<'a>(
-        files: impl IntoIterator<Item = (&'a str, &'a str)>,
-    ) -> anyhow::Result<Self> {
-        let mut case = Self::new();
-        case.write_files(files)?;
-        Ok(case)
-    }
-
-    pub fn with_file(path: impl AsRef<Path>, content: &str) -> anyhow::Result<Self> {
-        let mut case = Self::new();
-        case.write_file(path, content)?;
-        Ok(case)
-    }
-
-    pub fn write_files<'a>(
-        &mut self,
-        files: impl IntoIterator<Item = (&'a str, &'a str)>,
-    ) -> anyhow::Result<()> {
-        for (path, content) in files {
-            self.write_file(path, content)?;
-        }
-
-        Ok(())
-    }
-
-    pub fn write_file(&mut self, path: impl AsRef<Path>, content: &str) -> anyhow::Result<()> {
-        self.test_env.write_file(path, content)
-    }
-
-    #[must_use]
-    pub fn command(&self) -> Command {
-        let mut command = Command::new(self.karva_bin());
-        command.current_dir(self.test_env.cwd()).arg("test");
-        command
-    }
-
-    #[must_use]
-    pub fn command_with_args(&self, args: &[&str]) -> Command {
-        let mut command = self.command();
-        command.args(args);
-        command
     }
 }
