+++ crates/karva_test/src/utils.rs
@@ -1,0 +1,38 @@
+use std::path::{Path, PathBuf};
+
+use anyhow::Context;
+
+/// Find the karva wheel in the target/wheels directory.
+/// Returns the path to the wheel file.
+pub fn find_karva_wheel() -> anyhow::Result<PathBuf> {
+    let karva_root = Path::new(env!("CARGO_MANIFEST_DIR"))
+        .parent()
+        .and_then(|p| p.parent())
+        .ok_or_else(|| anyhow::anyhow!("Could not determine KARVA_ROOT"))?
+        .to_path_buf();
+
+    let wheels_dir = karva_root.join("target").join("wheels");
+
+    let entries = std::fs::read_dir(&wheels_dir)
+        .with_context(|| format!("Could not read wheels directory: {}", wheels_dir.display()))?;
+
+    for entry in entries {
+        let entry = entry?;
+        let file_name = entry.file_name();
+        if let Some(name) = file_name.to_str() {
+            if name.starts_with("karva-")
+                && std::path::Path::new(name)
+                    .extension()
+                    .is_some_and(|ext| ext.eq_ignore_ascii_case("whl"))
+            {
+                return Ok(entry.path());
+            }
+        }
+    }
+
+    anyhow::bail!("Could not find karva wheel in target/wheels directory");
+}
+
+pub fn tempdir_filter(path: &Path) -> String {
+    format!(r"{}\\?/?", regex::escape(path.to_str().unwrap()))
+}
